# src/main/resources/openapi/openapi.yml
openapi: 3.0.0
info:
  title: MktplaceApp APIs
  version: 1.0.0
  description: APIs for the marketplace application for buying and selling gaming assets
servers:
  - url: http://localhost:8080/api/v1
paths:
  /users/{userId}:
    get:
      operationId: getUserById
      summary: Get user details by Id
      tags:
        - User
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user to retrieve
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
        '500':
          description: Internal server error
  /users:
    post:
      operationId: createUser
      summary: Create a new user
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreationRequest'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreationResponse'
        '500':
          description: Internal server error
  /users/{userId}/products:
    get:
      operationId: getAllProductsListedByUser
      summary: Get all products listed by a user
      tags:
        - User
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user to retrieve product listings for
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Internal server error
  /users/{userId}/bookings:
    get:
      operationId: getAllBookingsByUser
      summary: Get all bookings done by a user / all products bought by a user
      tags:
        - User
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user to retrieve bookings for
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '500':
          description: Internal server error
...
  /paycallback:
    post:
      operationId: paycallback
      summary: API for payment callback
      tags:
        - Payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCallbackRequest'
      responses:
        '200':
          description: payment sttaus moved to success / failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentCallbackResponse'
        '500':
          description: Internal server error
components:
  schemas:
    UserCreationRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 20
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)[a-zA-Z\\d]{8,}$"
      required:
        - username
        - email
        - password
    UserCreationResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - success
            - failure
        user:
          $ref: '#/components/schemas/User'
          description: The created user object
...
    PaymentCallbackRequest:
      type: object
      properties:
        payment_id:
          type: integer
          format: int64
        booking_id:
          type: integer
          format: int64
        transaction_amount:
          type: number
          format: double
        status:
          type: string
          enum:
            - initiated
            - success
            - failed
            - reversed
            - refunded
      required:
        - payment_id
        - booking_id
        - transaction_amount
        - status
    PaymentCallbackResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - success
            - failure
        payment:
          $ref: '#/components/schemas/Payment'
          description: The created payment object with updated status
    User:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
        username:
          type: string
          minLength: 1  # Minimum length of username
          maxLength: 50 # Maximum length of username
        email:
          type: string
          format: email  # Format validation for email
        password:
          type: string
          format: password
          readOnly: true
          minLength: 8   # Minimum length of password
          maxLength: 20  # Maximum length of password
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)[a-zA-Z\\d]{8,}$"  # Password pattern (at least one lowercase, one uppercase, one digit)
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        last_purchase_timestamp:
          type: integer
          format: int64
    Product:
      type: object
      properties:
        product_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 500
        base_price:
          type: number
          format: double
        quantity_available:
          type: integer
          format: int64
        quantity_sold:
          type: integer
          format: int64
        quantity_on_hold:
          type: integer
          format: int64
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
    Booking:
      type: object
      properties:
        booking_id:
          type: integer
          format: int64
        buyer_id:
          type: integer
          format: int64
        seller_id:
          type: integer
          format: int64
        product_id:
          type: integer
          format: int64
        transaction_amount:
          type: number
          format: double
        payment_id:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - created
            - hold
            - booked
            - cancelled
            - refunded
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
    Payment:
      type: object
      properties:
        payment_id:
          type: integer
          format: int64
        booking_id:
          type: integer
          format: int64
        transaction_amount:
          type: number
          format: double
        status:
          type: string
          enum:
            - initiated
            - success
            - failed
            - reversed
            - refunded
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64